Fix building on iOS
git-svn-id: svn://srv-dc/svn2/Citrus@193176 20bf7d6a-bd21-9d4d-a3c4-a2c14bb8c68a

+++ Lime/Source/Graphics/OpenGL/RenderBatch.cs
@@ -71,7 +71,7 @@ namespace Lime
 			} else {
 				IndexBuffer.Bind();
 			}
-			GL.DrawElements(PrimitiveType.Triangles, IndexBuffer.IndexCount, DrawElementsType.UnsignedShort, offset);
+			GL.DrawElements(PrimitiveType.Triangles, IndexBuffer.IndexCount, DrawElementsType.UnsignedShort, (IntPtr)offset);
 			Renderer.DrawCalls++;
 		}
